Fix cross DC connecitons issue
In RequestRouter, we maintain a connection pool for all connected edc hosts. Connections
in this pool are lazy loaded. However, 10 ms timeout caused lots of cross continent connection
time out or stuck. Lack of protection on this connection pool, a lot of client threads would try
to create tens of connections to the same edc host at the same time and time out, which cause
a large number of dead threads.

Signed-off-by: Jin-cheng Chen <Jin-cheng.Chen@turn.com>